Treba uraditi korpu

+++ src/org/bildit/model/dto/Phone.java
@@ -7,7 +7,7 @@ public class Phone {
 	private String model;
 	private String operating_system;
 	private String description;
-	private double price;
+	private int price;
 	private int amount;
 	private String picture;
 	
@@ -55,11 +55,11 @@ public class Phone {
 		this.description = description;
 	}
 
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
